[fuchsia.settings.policy] Drop legacy support for VolumePolicyController
This removes `legacy=true` on the VolumePolicyController protocol, which
was removed in I7c9628c549ac3bb902883e131b7e3198aea02810. It is no
longer available in any supported API level.

The motivation for this is to line up legacy=true with the current set
of supported API levels for the implementation of RFC-0232.

Bug: 42085274
Change-Id: I648ee1030cbf3f77b20309f44217a1b787390c66
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/996532
Commit-Queue: Mitchell Kember <mkember@google.com>
API-Review: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Chip Fukuhara <cfukuhara@google.com>