[submodules] Add test submodule
... and an associated shac check that just prints the contents of the
submodule's README to confirm whether the submodule was checked out and
patched correctly.

This will be used to test submodules-as-source-of-truth workflows - all
previously existing submodules are still synced from Jiri manifests,
this is the first one that doesn't have a corresponding Jiri manifest
entry.

Bug: 326595281
Change-Id: Id39b35143c15f3e13b9bcdadbddb46214b4e3d78
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/996492
Reviewed-by: Danielle Kay <danikay@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>