[fxfs] Use fixed-with ints for checksums
Stores checksums as an array of little-endian integers in a  Vec<u8>
with some custom zero-copy accessors to &[Checksum].  This bypasses
bincode's preference of using varint encoding for u64 types by default,
which for checksums is wasteful (9 bytes instead of 8) as well as quite
a bit slower to deserialize.

The zero-copy accessors require little-endianness, so I've asserted that
we are building Fxfs on that for now since it's unlikely to change.

I also found that the journal wasn't checking extents in the root store
for validity if they didn't have checksums, which I corrected (and
removed two fsck test cases that would catch this otherwise -- now the
journal simply ignores the malformed extents during replay).

Bug: 323571978
Change-Id: I053ebdac87940325da36e0c8a867e2ec11d3f194
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/996156
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>