[fidl][rust] Do not call get_unchecked on empty slices
This fixes a bug in FIDL Rust decoding that is failing assertions in the
canary builders that roll the Rust toolchain.

For certain types, such as vector<uint8>, we optimize decoding to a
memcpy. We were using `get_unchecked` to obtain the source pointer in
the buffer. This was incorrect for empty vectors because `get_unchecked`
returns a reference, not a pointer.

I made two changes. Either one on its own would fix the problem, but
they are both worthwhile:

1. Use slice.as_ptr().add(offset) instead of slice.get_unchecked(offset)
   before calling memcpy. This avoids creating a reference at all.

2. Assert that len > 0 in decoder.out_of_line_offset, as we already do
   in encoder.out_of_line_offset. This significantly simplifies that
   method's `unsafe` rationale comment. It only required adding len != 0
   checks to a few callers (most pass inline sizes that cannot be zero).

Test: fx test fidl_rust_conformance_tests
Fixed: 327029914
Change-Id: Ic3cb0c6b69f5c061ca1bc564e8bf63d1b3314ed7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/997122
Reviewed-by: David Koloski <dkoloski@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Fuchsia-Auto-Submit: Mitchell Kember <mkember@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>