[dhcp-client] Annotate logs with interface
Improves debug logging of the out-of-stack DHCPv4 client by annotating
each log line with the interface ID the client is attached to.
Also downgrades some WARN logs to DEBUG that are misleading
diagnostically when logged at WARN (e.g. a UDP message on the "wrong
port" is just an incoming non-DHCP packet and not cause for suspicion).

Test: dhcp_client_core, dhcp-client-tests
Change-Id: I2e99dbc6f8d9ae602f0ce7511ddf4777211605d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/997253
Commit-Queue: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>