[fake-clock] Disallow concurrent use of fake and UTC clocks
Instrument the `fake-clock` library to disallow by default
the use of the monotonic clock and the UTC clock at the
same time. This situation could be a source of subtle
bugs, so we force the test binary to crash if this happens
without the dev realizing it.

The developer can make a choice of how to resolve this - either
use the proper test fixture, or knowingly disregard the issue.
The choice is documented in the README.md.

The fake-clock library is today used in rare spots in Fuchsia.
Those are quite low-level and do not care about the UTC
clock. This means we can firm up the behavior around clock use
without affecting current clients.

Tests:
  I verified manually that the death test in fact performs as
  you'd expect. The only issue is that the test fixture is not
  set up correctly. I would still like to have this change in
  to unblock engineers wanting to use fake-clock, and will
  go back to figuring out the correct incantation to enable
  the death test again.

Bug: b/327476961
Change-Id: Idee145050a75844be22ec21676f556c2e34edef4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/999777
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>