[wlan] Prefer async fuchsia::test over constructing an executor
In many places there was an implicit requirement to construct an
executor in tests to construct fakes properly. Now, constructors for
fakes are async functions documented with an explanation for why
they're async. As usual, to call these async constructor, tests are
free to construct an executor explicitly or be written in an async
function so they can await calls to async functions.

Before this change, various functions required an unused
`fuchsia_async::TestExecutor` argument to ensure an executor was set
up before constructing the faked value. This was implicitly required
because some parts of wlan-sme and wlan-mlme may call
`wlan_common::timer::Timer::now` which is not an async function but
requires an executor has been set up (https://fxbug.dev/327499461).

This change makes `Timer::now` an unsafe function as a lint to remind
the caller of the additional requirement that it must be called after
an executor has been set up. And many, if not all, functions that
previously required an unused `fuchsia_async::TestExecutor` argument
are now async themselves. As a result, tests that construct fakes
will do so through an async constructor for the fake.

The motivation for this change is to make the use of async in
wlan-mlme tests more ergonomic as I have a tentative plan to plumb
async further down in wlan-mlme.

Test: fx test wlan-mlme-tests wlan-sme-test wlansoftmac-rust-tests
Bug: 327499461
Change-Id: I0aa59dfd62f5d5e089534f763a563afe22612577
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/999699
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kiet Tran <kiettran@google.com>
Commit-Queue: Charles Celerier <chcl@google.com>