[debugger] Format ELF symbols with file names.
In some unusual cases you can have an ELF symbol with a file name. This
can happen if the DIE information is corrupt or unloadable, or if the
DIEs were omitted from the binary (when compiling with "line info
only").

Previously the separator was missing between the ELF symbol name and the
file name.

Change-Id: I05a417213cd65ee5761c1f1b49f7c1824ad82ae5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1001366
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>