[mutex_pi_exerciser] Convert to using the zx_profile system calls
This CL changes the mutex_pi_exerciser to use the `zx_profile_create`
system call to create profile objects instead of going through the
ProfileProvider. This is part of the work to deprecate the
ProfileProvider.

The alternative of using the RoleManager (the ProfileProvider's
replacement) was considered, but that API is a bit too rigid for
the kinds of desk testing mutex_pi_exerciser runs.

Bug: 323262398
Change-Id: I06a4c86046bb40d859e6d142655fa6e249d8b49c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1001575
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>