[bt][hfp-hf] Translate +CIEVs to typed values.
The AG sends +CIEV indicators with an integer index and value.  The
mapping of these indices to specific indicators is originally specified
in a +CIND sent as part of SLCI.  This change adds a module to store
these indices and map them to typed AG indicator enums. It also adds a
ProcedureOutput variant to set the initial values of these indicators.

Glue code to use this module and ProcedureOutput variant is still to
come.

Test: fx test bt-hfp-hands-free-tests

Change-Id: I3e422bc1d413eafba16511ba94513d7c4e2235f7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/895561
Reviewed-by: Marie Janssen <jamuraa@google.com>
Commit-Queue: Matthew Kehrt <kehrt@google.com>