[display][controller] make image-related FIDL requests
This change provides a single internally stored `fdf::Arena arena_`,
created upon construction, for use for FIDL calls in the DFV2 style.

Driver::ImportImage is not implemented in this change because we need to
design a new image-related data structure in the api-types-cpp library
first.

Bug: b/314126494
Test: fx test api-types-cpp-tests display-coordinator-unittests

Change-Id: I6b20ffd602db9ec70c48a5cfb87735c19ba60abc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/991773
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Caroline Liu <carolineliu@google.com>