[perf_publish] Fix failure that occurs when unit test is run from a branch
When tests are run in Fuchsia Infra on branches, they are run with the
"RELEASE_VERSION" env var set.

One of the perf_publish unit test cases
(test_run_converter_accept_expected_metrics) has the problem that it
fails when that env var is set, such as when run like this:

  RELEASE_VERSION=foo fx test host_x64/obj/src/performance/lib/perf_publish/tests/perf_publish_test.sh

In that case, the perf_publish library adds a "--product-versions"
argument to the catapult_converter command line, which doesn't match
what the unit test expects.

Fix by overriding the environment variable dict that is used in this
test.

Bug: b/328272533
Test: see above
Change-Id: Ifa62a1e4850488a4cf8a4e569ab93634d0c0c67a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1002292
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>