[idk] Remove dead rust code.
//build/sdk/meta contains a rust crate full of types that model the IDK
metadata. While these are nicer and more strongly typed than the
python code that manipulates the IDK, there's a problem: they're not
really used.

In fxrev.dev/998794, chok@ was under the reasonable (but incorrect),
impression that adding a new type of atom to this library was both
necessary and sufficient for adding a new type of thing to the IDK, but
unfortunately it's neither. Clearly, this code is doing more harm than
good.

Some of this library is used, specifically: code to do with product
bundles, and code that ffx uses to locate binaries.

This deletes everything else. RIP.

Change-Id: I3cdf8660b7f74fc393b323389f0499a1f6a06869
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1001513
Commit-Queue: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: David Turner <digit@google.com>