[package-tool] Support api-level=HEAD.
More generally, this adds support for the mutable "pseudo-API-levels"
HEAD, LEGACY, and eventually NEXT.

The ABI revision and status assigned to these API levels is not correct
- it should use a random "per release" ABI revision (see
fxbug.dev/42082683). But, we're getting there one step at a time. We
first need to make sure that everyone actually specifies which API Level
they're targeting, always.

This also changes the platform build to specify `LEGACY` while building
the platform, and a specific API level in IDK subbuilds.

The fact that the addition of these API levels is implemented in
`parse_version_history` is perhaps a bit strange. It could come earlier,
with `HEAD` and `LEGACY` actually appearing in `version_history.json`,
but the ABI revision will eventually be dynamically generated, so they'd
still need to be treated specially there. They could also be added
later, but then every method in VersionHistory would need to treat them
specially, since they wouldn't appear in the constant.

We may need to reconsider this, especially as we explore what it means
for SDKs to support targeting `NEXT` and `HEAD`, but I think this will
do for now.

Also make the error messages marginally prettier.

Fixed: 327656572
Bug: 327657448
Fixed: 326655565
Bug: 305961460

Change-Id: I987f768cc7954a17bd3a5dedfcffacd917269875
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/998738
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: David Dorwin <ddorwin@google.com>
Commit-Queue: Hunter Freyer <hjfreyer@google.com>